04-spectral-reconstruction/requirements.txt
h5py
ipykernel
colour-science
matplotlib
numpy
fastapi
uvicorn
torch
einops
torchvision
torchmetrics
tqdm
scikit-learn
